Refactor 'classify_paths' (cut never-used diff_hsum)
'classify_paths' is called two times in 'compute_moves'.

In the first call, files don't have actual hashes,
they only may have unique file sizes in .hashsum attribute.
So rare cases of the same path, same size and different hashes
are not checked, not added to 'modified'.
(And of course it is too expensive to check for them).

In the second call, they don't have hashsum,
so they are skipped again.